pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Katrin Torsunova <katrin.tep@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
pytest = "8.0.2"
allure-pytest = "2.13.2"
pytest-check = "2.2.0"
requests = "^2.31.0"
curlify = "2.2.1"
urllib3 = "2.2.1"
graphql-query = "1.2.2"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"